+++ src/hooks/useGTMEvent.ts
@@ -6,7 +6,6 @@ type GTMEvent = {
 const useGTMEvent = () => {
     return (eventData: GTMEvent) => {
         if (typeof window !== "undefined" && window.dataLayer) {
-            console.log(eventData);
             window.dataLayer.push(eventData);
         }
     };
